requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2
email-validator==2.1.0.post1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.27
Werkzeug==3.0.1
WTForms==3.1.2 